web/requirements.txt
djangorestframework
drfdocs